+++ spec/requests/members/destroy_spec.rb
@@ -1,10 +1,8 @@
 require 'rails_helper'
 
 RSpec.describe 'Members', type: :request do
-  include_context 'リクエストスペース作成'
-
-  # PATCH/PUT /members/:customer_code/:user_code（ベースドメイン） メンバー権限変更(処理)
-  # PATCH/PUT /members/:customer_code/:user_code.json（ベースドメイン） メンバー権限変更API
+  # DELETE /members/:customer_code/:user_code（ベースドメイン） メンバー解除(処理)
+  # DELETE /members/:customer_code/:user_code.json（ベースドメイン） メンバー解除API
   # 前提条件
   #   なし
   # テストパターン
@@ -13,7 +11,8 @@ RSpec.describe 'Members', type: :request do
   #   顧客: 所属, 未所属, 存在しない, ない → 事前にデータ作成
   #   対象: ない, 自分, Owner, Admin, Member → 事前にデータ作成
   #   ベースドメイン, 存在するサブドメイン, 存在しないサブドメイン → 事前にデータ作成
-  describe 'PATCH /destroy' do
+  describe 'DELETE /destroy' do
+    include_context 'リクエストスペース作成'
     include_context 'メンバー作成', 1, 1, 1, 0, 'ASC'
     include_context 'メンバー作成（対象外）', 'ASC'