+++ spec/requests/spaces/new_spec.rb
@@ -8,6 +8,8 @@ RSpec.describe 'Spaces', type: :request do
   #   未ログイン, ログイン中, ログイン中（削除予約済み） → データ＆状態作成
   #   ベースドメイン, 存在するサブドメイン, 存在しないサブドメイン → 事前にデータ作成
   describe 'GET /new' do
+    include_context 'リクエストスペース作成'
+
     # テスト内容
     shared_examples_for 'ToOK' do
       it '成功ステータス' do