+++ spec/views/spaces/edit.html.erb_spec.rb
@@ -2,13 +2,26 @@ require 'rails_helper'
 
 RSpec.describe 'spaces/edit', type: :view do
   let!(:customer) { FactoryBot.create(:customer) }
-  before { @space = assign(:space, FactoryBot.create(:space, customer_id: customer.id)) }
+  before do
+    @space = assign(:space, FactoryBot.create(:space, customer_id: customer.id))
+    @request_space = @space
+  end
 
   it 'renders the edit space form' do
     render
     assert_select 'form[action=?][method=?]', update_space_path, 'post' do
       assert_select 'input[name=?]', 'space[subdomain]'
+      assert_select 'input[name=?]', 'space[public_flag]'
       assert_select 'input[name=?]', 'space[name]'
+      assert_select 'input[name=?]', 'space[purpose]'
+      assert_select 'input[name=?]', 'commit'
+    end
+  end
+
+  it 'renders the edit space form' do
+    render
+    assert_select 'form[action=?][method=?]', update_space_image_path, 'post' do
+      assert_select 'input[name=?]', 'space[image]'
       assert_select 'input[name=?]', 'commit'
     end
   end